pyproject.toml
[tool.poetry]
name = "wormhole-ui"
version = "0.1.0"
description = "UI for Magic Wormhole - get things from one computer to another safely"
authors = ["sneakypete81 <sneakypete81@gmail.com>"]

[tool.poetry.scripts]
wormhole-ui = 'wormhole_ui.main:run'

[tool.poetry.dependencies]
python = "^3.5, <3.9"
magic_wormhole = "^0.11.2"
PySide2 = "^5.14.0"
qt5reactor = "^0.6"
humanize = "^0.5.1"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
pyinstaller = "^3.5"
pywin32-ctypes = { version = "^0.2.0", platform = "win32" }
pefile = { version = "^2019.4.18", platform = "win32" }
macholib = { version = "^1.13", platform = "darwin" }
flake8 = "^3.7.9"
black = { version = "^19.10b0", python = "^3.6" }

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
